Split component managers out from syntax definition.
