+++ packages/ember-glimmer/lib/component-managers/abstract.js
