+++ packages/ember-glimmer/lib/renderer.js
@@ -12,8 +12,8 @@ import {
   getViewId
 } from 'ember-views';
 import { BOUNDS } from './component';
-import { RootComponentDefinition } from './syntax/curly-component';
-import { TopLevelOutletComponentDefinition } from './syntax/outlet';
+import { RootComponentDefinition } from './component-managers/root';
+import { TopLevelOutletComponentDefinition } from './component-managers/outlet';
 import { assert } from 'ember-debug';
 
 const { backburner } = run;
