+++ packages/ember-glimmer/lib/syntax/outlet.js
@@ -1,21 +1,61 @@
-/**
-@module ember
-@submodule ember-glimmer
-*/
-import { generateGuid, guidFor } from 'ember-utils';
-import {
-  ComponentDefinition,
-  CompiledArgs
-} from '@glimmer/runtime';
-import { DEBUG } from 'ember-env-flags';
-import { _instrumentStart } from 'ember-metal';
-import { RootReference } from '../utils/references';
-import AbstractManager from './abstract-manager';
 import {
   UpdatableTag,
   ConstReference,
   combine
 } from '@glimmer/reference';
+import { OutletComponentDefinition } from '../component-managers/outlet';
+
+class OutletComponentReference {
+  constructor(outletNameRef, parentOutletStateRef) {
+    this.outletNameRef = outletNameRef;
+    this.parentOutletStateRef = parentOutletStateRef;
+    this.definition = null;
+    this.lastState = null;
+    let outletStateTag = this.outletStateTag = new UpdatableTag(parentOutletStateRef.tag);
+    this.tag = combine([outletStateTag.tag, outletNameRef.tag]);
+  }
+
+  value() {
+    let { outletNameRef, parentOutletStateRef, definition, lastState } = this;
+
+    let outletName = outletNameRef.value();
+    let outletStateRef = parentOutletStateRef.get('outlets').get(outletName);
+    let newState = this.lastState = outletStateRef.value();
+
+    this.outletStateTag.update(outletStateRef.tag);
+
+    definition = revalidate(definition, lastState, newState);
+
+    let hasTemplate = newState && newState.render.template;
+
+    if (definition) {
+      return definition;
+    } else if (hasTemplate) {
+      return this.definition = new OutletComponentDefinition(outletName, newState.render.template);
+    } else {
+      return this.definition = null;
+    }
+  }
+}
+
+function revalidate(definition, lastState, newState) {
+  if (!lastState && !newState) {
+    return definition;
+  }
+
+  if (!lastState && newState || lastState && !newState) {
+    return null;
+  }
+
+  if (
+    newState.render.template === lastState.render.template &&
+    newState.render.controller === lastState.render.controller
+  ) {
+    return definition;
+  }
+
+  return null;
+}
 
 function outletComponentFor(vm, args) {
   let { outletState } = vm.dynamicScope();
@@ -88,186 +128,3 @@ export function outletMacro(name, params, hash, builder) {
   builder.component.dynamic(definitionArgs, outletComponentFor, emptyArgs);
   return true;
 }
-
-class OutletComponentReference {
-  constructor(outletNameRef, parentOutletStateRef) {
-    this.outletNameRef = outletNameRef;
-    this.parentOutletStateRef = parentOutletStateRef;
-    this.definition = null;
-    this.lastState = null;
-    let outletStateTag = this.outletStateTag = new UpdatableTag(parentOutletStateRef.tag);
-    this.tag = combine([outletStateTag.tag, outletNameRef.tag]);
-  }
-
-  value() {
-    let { outletNameRef, parentOutletStateRef, definition, lastState } = this;
-
-    let outletName = outletNameRef.value();
-    let outletStateRef = parentOutletStateRef.get('outlets').get(outletName);
-    let newState = this.lastState = outletStateRef.value();
-
-    this.outletStateTag.update(outletStateRef.tag);
-
-    definition = revalidate(definition, lastState, newState);
-
-    let hasTemplate = newState && newState.render.template;
-
-    if (definition) {
-      return definition;
-    } else if (hasTemplate) {
-      return this.definition = new OutletComponentDefinition(outletName, newState.render.template);
-    } else {
-      return this.definition = null;
-    }
-  }
-}
-
-function revalidate(definition, lastState, newState) {
-  if (!lastState && !newState) {
-    return definition;
-  }
-
-  if (!lastState && newState || lastState && !newState) {
-    return null;
-  }
-
-  if (
-    newState.render.template === lastState.render.template &&
-    newState.render.controller === lastState.render.controller
-  ) {
-    return definition;
-  }
-
-  return null;
-}
-
-function instrumentationPayload({ render: { name, outlet } }) {
-  return { object: `${name}:${outlet}` };
-}
-
-function NOOP() {}
-
-class StateBucket {
-  constructor(outletState) {
-    this.outletState = outletState;
-    this.instrument();
-  }
-
-  instrument() {
-    this.finalizer = _instrumentStart('render.outlet', instrumentationPayload, this.outletState);
-  }
-
-  finalize() {
-    let { finalizer } = this;
-    finalizer();
-    this.finalizer = NOOP;
-  }
-}
-
-class OutletComponentManager extends AbstractManager {
-  prepareArgs(definition, args) {
-    return null;
-  }
-
-  create(environment, definition, args, dynamicScope) {
-    if (DEBUG) {
-      this._pushToDebugStack(`template:${definition.template.meta.moduleName}`, environment);
-    }
-
-    let outletStateReference = dynamicScope.outletState = dynamicScope.outletState.get('outlets').get(definition.outletName);
-    let outletState = outletStateReference.value();
-    return new StateBucket(outletState);
-  }
-
-  layoutFor(definition, bucket, env) {
-    return env.getCompiledBlock(OutletLayoutCompiler, definition.template);
-  }
-
-  getSelf({ outletState }) {
-    return new RootReference(outletState.render.controller);
-  }
-
-  getTag() {
-    return null;
-  }
-
-  getDestructor() {
-    return null;
-  }
-
-  didRenderLayout(bucket) {
-    bucket.finalize();
-
-    if (DEBUG) {
-      this.debugStack.pop();
-    }
-  }
-
-  didCreateElement() {}
-  didCreate(state) {}
-  update(bucket) {}
-  didUpdateLayout(bucket) {}
-  didUpdate(state) {}
-}
-
-const MANAGER = new OutletComponentManager();
-
-class TopLevelOutletComponentManager extends OutletComponentManager {
-  create(environment, definition, args, dynamicScope) {
-    if (DEBUG) {
-      this._pushToDebugStack(`template:${definition.template.meta.moduleName}`, environment);
-    }
-    return new StateBucket(dynamicScope.outletState.value());
-  }
-
-  layoutFor(definition, bucket, env) {
-    return env.getCompiledBlock(TopLevelOutletLayoutCompiler, definition.template);
-  }
-}
-
-const TOP_LEVEL_MANAGER = new TopLevelOutletComponentManager();
-
-
-export class TopLevelOutletComponentDefinition extends ComponentDefinition {
-  constructor(instance) {
-    super('outlet', TOP_LEVEL_MANAGER, instance);
-    this.template = instance.template;
-    generateGuid(this);
-  }
-}
-
-class TopLevelOutletLayoutCompiler {
-  constructor(template) {
-    this.template = template;
-  }
-
-  compile(builder) {
-    builder.wrapLayout(this.template);
-    builder.tag.static('div');
-    builder.attrs.static('id', guidFor(this));
-    builder.attrs.static('class', 'ember-view');
-  }
-}
-
-TopLevelOutletLayoutCompiler.id = 'top-level-outlet';
-
-class OutletComponentDefinition extends ComponentDefinition {
-  constructor(outletName, template) {
-    super('outlet', MANAGER, null);
-    this.outletName = outletName;
-    this.template = template;
-    generateGuid(this);
-  }
-}
-
-export class OutletLayoutCompiler {
-  constructor(template) {
-    this.template = template;
-  }
-
-  compile(builder) {
-    builder.wrapLayout(this.template);
-  }
-}
-
-OutletLayoutCompiler.id = 'outlet';
