+++ packages/ember-glimmer/lib/utils/curly-component-state-bucket.js
